Dockerization - first commit
-Added dockerfile images for each microservice
-Added docker-compose.yml file for container inits

To do:
Logging out with spring security does not work. It redirects to the hostname that is client container's id instead of localhost and container's internal docker port instead of gateway port. Even if SecurityConfig's logout success url is set to http://localhost:8080/client/login redirection works, but user is still not logged out